refactor: rename components to singular 'user' for consistency
